Move Alarm handling out of TableRepairJob
* Create hook in update called postUpdate
* Moved Alarm handling out of TableRepairJob
* Allows creation of TableRepairJob's without Alarms
