+++ core/src/test/java/com/ericsson/bss/cassandra/ecchronos/core/repair/TestTableRepairJob.java
@@ -91,9 +91,6 @@ public class TestTableRepairJob
     @Mock
     private TableRepairMetrics myTableRepairMetrics;
 
-    @Mock
-    private RepairFaultReporter myFaultReporter;
-
     private TableRepairJob myRepairJob;
 
     private MockedClock myClock = new MockedClock();
@@ -128,13 +125,10 @@ public class TestTableRepairJob
                 .withTableReference(myTableReference)
                 .withJmxProxyFactory(myJmxProxyFactory)
                 .withRepairState(myRepairState)
-                .withFaultReporter(myFaultReporter)
                 .withTableRepairMetrics(myTableRepairMetrics)
                 .withRepairConfiguration(repairConfiguration)
                 .withRepairLockType(RepairLockType.VNODE)
                 .build();
-
-        myRepairJob.setClock(myClock);
     }
 
     @After
@@ -276,220 +270,5 @@ public class TestTableRepairJob
         myRepairJob.postExecute(true);
 
         assertThat(myRepairJob.getLastSuccessfulRun()).isEqualTo(lastRun);
-    }
-
-    @Test
-    public void testThatWarningAlarmIsSentAndCeased()
-    {
-        // setup - not repaired
-        long daysSinceLastRepair = 2;
-        long start = System.currentTimeMillis();
-        long lastRepaired = start - TimeUnit.DAYS.toMillis(daysSinceLastRepair);
-
-        Map<String, Object> expectedData = new HashMap<>();
-        expectedData.put(RepairFaultReporter.FAULT_KEYSPACE, keyspaceName);
-        expectedData.put(RepairFaultReporter.FAULT_TABLE, tableName);
-
-        // mock - not repaired
-        doReturn(lastRepaired).when(myRepairStateSnapshot).lastRepairedAt();
-        doReturn(true).when(myRepairStateSnapshot).canRepair();
-        myClock.setTime(start);
-
-        assertThat(myRepairJob.runnable()).isTrue();
-
-        // verify - not repaired
-        verify(myFaultReporter).raise(eq(RepairFaultReporter.FaultCode.REPAIR_WARNING), eq(expectedData));
-
-        // setup - repaired
-        lastRepaired = start;
-        start = System.currentTimeMillis();
-
-        // mock - repaired
-        doReturn(lastRepaired).when(myRepairStateSnapshot).lastRepairedAt();
-        doReturn(false).when(myRepairStateSnapshot).canRepair();
-        myClock.setTime(start);
-
-        myRepairJob.runnable();
-
-        // verify alarm ceased in preValidate
-        verify(myFaultReporter).cease(eq(RepairFaultReporter.FaultCode.REPAIR_WARNING), eq(expectedData));
-        reset(myFaultReporter);
-
-        myRepairJob.postExecute(true);
-
-        // verify - repaired
-        verify(myFaultReporter).cease(eq(RepairFaultReporter.FaultCode.REPAIR_WARNING), eq(expectedData));
-    }
-
-    @Test
-    public void testThatWarningAlarmIsSentAndCeasedExternalRepair()
-    {
-        // setup - not repaired
-        long daysSinceLastRepair = 2;
-        long start = System.currentTimeMillis();
-        long lastRepaired = start - TimeUnit.DAYS.toMillis(daysSinceLastRepair);
-
-        Map<String, Object> expectedData = new HashMap<>();
-        expectedData.put(RepairFaultReporter.FAULT_KEYSPACE, keyspaceName);
-        expectedData.put(RepairFaultReporter.FAULT_TABLE, tableName);
-
-        // mock - not repaired
-        doReturn(lastRepaired).when(myRepairStateSnapshot).lastRepairedAt();
-        doReturn(true).when(myRepairStateSnapshot).canRepair();
-        myClock.setTime(start);
-
-        assertThat(myRepairJob.runnable()).isTrue();
-
-        // verify - not repaired
-        verify(myFaultReporter).raise(eq(RepairFaultReporter.FaultCode.REPAIR_WARNING), eq(expectedData));
-
-        // setup - repaired
-        lastRepaired = start;
-        start = System.currentTimeMillis();
-
-        // mock - repaired
-        doReturn(lastRepaired).when(myRepairStateSnapshot).lastRepairedAt();
-        doReturn(false).when(myRepairStateSnapshot).canRepair();
-        myClock.setTime(start);
-
-        assertThat(myRepairJob.runnable()).isFalse();
-
-        // verify - repaired
-        verify(myFaultReporter).cease(eq(RepairFaultReporter.FaultCode.REPAIR_WARNING), eq(expectedData));
-    }
-
-    @Test
-    public void testThatErrorAlarmIsSentAndCeased()
-    {
-        // setup - not repaired
-        long daysSinceLastRepair = GC_GRACE_DAYS;
-        long start = System.currentTimeMillis();
-        long lastRepaired = start - TimeUnit.DAYS.toMillis(daysSinceLastRepair);
-
-        Map<String, Object> expectedData = new HashMap<>();
-        expectedData.put(RepairFaultReporter.FAULT_KEYSPACE, keyspaceName);
-        expectedData.put(RepairFaultReporter.FAULT_TABLE, tableName);
-
-        // mock - not repaired
-        doReturn(lastRepaired).when(myRepairStateSnapshot).lastRepairedAt();
-        doReturn(true).when(myRepairStateSnapshot).canRepair();
-        myClock.setTime(start);
-
-        assertThat(myRepairJob.runnable()).isTrue();
-
-        // verify - not repaired
-        verify(myFaultReporter).raise(eq(RepairFaultReporter.FaultCode.REPAIR_ERROR), eq(expectedData));
-
-        // setup - repaired
-        lastRepaired = start;
-        start = System.currentTimeMillis();
-
-        // mock - repaired
-        doReturn(lastRepaired).when(myRepairStateSnapshot).lastRepairedAt();
-        doReturn(false).when(myRepairStateSnapshot).canRepair();
-        myClock.setTime(start);
-
-        myRepairJob.postExecute(true);
-
-        // verify - repaired
-        verify(myFaultReporter).cease(eq(FaultCode.REPAIR_WARNING), eq(expectedData));
-    }
-
-    @Test
-    public void testThatErrorAlarmIsSentAndCeasedExternalRepair()
-    {
-        // setup - not repaired
-        long daysSinceLastRepair = GC_GRACE_DAYS;
-        long start = System.currentTimeMillis();
-        long lastRepaired = start - TimeUnit.DAYS.toMillis(daysSinceLastRepair);
-
-        Map<String, Object> expectedData = new HashMap<>();
-        expectedData.put(RepairFaultReporter.FAULT_KEYSPACE, keyspaceName);
-        expectedData.put(RepairFaultReporter.FAULT_TABLE, tableName);
-
-        // mock - not repaired
-        doReturn(lastRepaired).when(myRepairStateSnapshot).lastRepairedAt();
-        doReturn(true).when(myRepairStateSnapshot).canRepair();
-        myClock.setTime(start);
-
-        assertThat(myRepairJob.runnable()).isTrue();
-
-        // verify - not repaired
-        verify(myFaultReporter).raise(eq(RepairFaultReporter.FaultCode.REPAIR_ERROR), eq(expectedData));
-
-        // setup - repaired
-        lastRepaired = start;
-        start = System.currentTimeMillis();
-
-        // mock - repaired
-        doReturn(lastRepaired).when(myRepairStateSnapshot).lastRepairedAt();
-        doReturn(false).when(myRepairStateSnapshot).canRepair();
-        myClock.setTime(start);
-
-        myRepairJob.runnable();
-
-        // verify - repaired
-        verify(myFaultReporter).cease(eq(RepairFaultReporter.FaultCode.REPAIR_WARNING), eq(expectedData));
-    }
-
-    @Test
-    public void testThatAlarmIsNotSentWhenGcGraceIsBelowRepairInterval()
-    {
-        // setup - not repaired
-        int tableGcGrace = (int) TimeUnit.HOURS.toSeconds(22);
-        long hoursSinceLastRepair = 23;
-        long start = System.currentTimeMillis();
-        long lastRepaired = start - TimeUnit.HOURS.toMillis(hoursSinceLastRepair);
-
-        // mock - not repaired
-        doReturn(tableGcGrace).when(myTableOptionsMetadata).getGcGraceInSeconds();
-        doReturn(lastRepaired).when(myRepairStateSnapshot).lastRepairedAt();
-        doReturn(true).when(myRepairStateSnapshot).canRepair();
-        myClock.setTime(start);
-
-        assertThat(myRepairJob.runnable()).isFalse();
-
-        // verify - not repaired
-        verify(myFaultReporter, never()).raise(any(RepairFaultReporter.FaultCode.class), anyMapOf(String.class, Object.class));
-    }
-
-    @Test
-    public void testLastSuccessfulRunIsBasedOnRepairHistory()
-    {
-        long timeOffset = TimeUnit.MINUTES.toMillis(1);
-        long now = System.currentTimeMillis() - TimeUnit.MINUTES.toMillis(2);
-        long lastRepairedAtWarning = now - TimeUnit.DAYS.toMillis(RUN_INTERVAL_IN_DAYS * 2);
-        long lastRepairedAtAfterRepair = now - TimeUnit.DAYS.toMillis(RUN_INTERVAL_IN_DAYS) + timeOffset;
-
-        myClock.setTime(now);
-
-        // We have waited 2 days to repair, send alarm and run repair
-        doReturn(lastRepairedAtWarning).when(myRepairStateSnapshot).lastRepairedAt();
-        doReturn(true).when(myRepairStateSnapshot).canRepair();
-
-        assertThat(myRepairJob.runnable()).isTrue();
-
-        assertThat(myRepairJob.getLastSuccessfulRun()).isEqualTo(lastRepairedAtWarning);
-        verify(myFaultReporter).raise(eq(FaultCode.REPAIR_WARNING), anyMapOf(String.class, Object.class));
-        verifyNoMoreInteractions(myFaultReporter);
-        reset(myFaultReporter);
-
-        // Repair has been completed
-        doReturn(lastRepairedAtAfterRepair).when(myRepairStateSnapshot).lastRepairedAt();
-        doReturn(false).when(myRepairStateSnapshot).canRepair();
-
-        myRepairJob.postExecute(true);
-
-        assertThat(myRepairJob.getLastSuccessfulRun()).isEqualTo(lastRepairedAtAfterRepair);
-        verify(myFaultReporter).cease(eq(FaultCode.REPAIR_WARNING), anyMapOf(String.class, Object.class));
-
-        // After 10 ms we can repair again
-        myClock.setTime(now + timeOffset);
-        doReturn(true).when(myRepairStateSnapshot).canRepair();
-
-        assertThat(myRepairJob.runnable()).isTrue();
-
-        assertThat(myRepairJob.getLastSuccessfulRun()).isEqualTo(lastRepairedAtAfterRepair);
-        verify(myFaultReporter, times(0)).raise(any(FaultCode.class), anyMapOf(String.class, Object.class));
     }
 }
